test: centralize legacy graph setup
Amp-Thread-ID: https://ampcode.com/threads/T-01a020d1-98ad-704c-84cb-274c975cec92
